refactor Files : File 정리, 필요없는 부분 삭제
