+++ app/src/main/java/com/dudoji/android/model/MapSectionManager.kt
@@ -1,9 +1,12 @@
-package com.dudoji.android.model.mapsection
+package com.dudoji.android.model
 
 import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.Color
 import android.graphics.Paint
+import com.dudoji.android.model.mapsection.BASIC_ZOOM_LEVEL
+import com.dudoji.android.model.mapsection.DetailedMapSection
+import com.dudoji.android.model.mapsection.MapSection
 import com.dudoji.android.util.mapsection.BitmapUtil
 import com.dudoji.android.util.mapsection.BitmapUtil.Companion.combineBitmapInGrid
 import com.dudoji.android.util.tile.TILE_SIZE
