+++ app/src/main/java/com/dudoji/android/model/RevealCircle.kt
@@ -1,3 +1,3 @@
-package com.dudoji.android.model.mapsection
+package com.dudoji.android.model
 
 data class RevealCircle(val lat: Double, val lng: Double, val radius: Double)
