+++ app/src/main/java/com/dudoji/android/util/location/IRevealCircleListener.kt
@@ -1,6 +1,6 @@
 package com.dudoji.android.util.location
 
-import com.dudoji.android.model.mapsection.RevealCircle
+import com.dudoji.android.model.RevealCircle
 import com.dudoji.android.util.listener.IListener
 
 interface IRevealCircleListener: IListener<RevealCircle> {
